pyproject.toml
[tool.poetry]
name = "fastapi-xml-parser"
version = "0.1.0"
description = ""
authors = ["topgun <reddevil.23031991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.5"}
asyncpg = "^0.30.0"
celery = "^5.4.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
flower = "^2.0.1"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
aiohttp = "^3.11.6"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
aiofiles = "^24.1.0"
alembic = "^1.14.0"
pydantic-settings = "^2.6.1"
pytest = "^8.3.3"
pytest-dotenv = "^0.5.2"
sqlalchemy-utils = "^0.41.2"
async-asgi-testclient = "^1.4.11"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
isort = "^5.13.2"
gunicorn = "^23.0.0"
uvicorn = "^0.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
